0.3.5 ALPHA-STABLE
Updated a few layout things.

Made the footer information and copyright a setting that you can change which pulls through.
Added a custom header block region for the course search block specifically. Will need to adapt this at a later stage.
Customised the login page.
Added a secondary brand colour to settings and made it apply to a few minor elements around the site.